refactor: cleaner structure and scheduling lib
