+++ calendar/Dockerfile
@@ -1,27 +1,17 @@
 # Build image gen app
-FROM golang:1.19-alpine as build_imagegen
+FROM golang:1.19-alpine as build
 
 WORKDIR /app
 
-COPY imagegen/go.mod imagegen/go.sum ./
+COPY go.mod go.sum ./
 
 RUN go mod download && go mod verify
 
-COPY imagegen/* .
+COPY * .
 
-RUN go build -v -o /app/calendar .
+RUN go build -v -o /app/calimage .
 
 RUN rm go.*
-
-# Build refresh server
-
-FROM golang:1.19-alpine as build_refresh
-
-WORKDIR /app
-
-COPY refreshserver/* .
-
-RUN go build -v -o /app/refreshserver .
 
 # Build image
 
@@ -29,12 +19,10 @@ FROM alpine:3.17
 
 WORKDIR /app
 
-COPY --from=build_imagegen /app/calendar .
-COPY --from=build_refresh /app/refreshserver .
-COPY ./entrypoint.sh .
+COPY --from=build /app/calimage .
 
 COPY config/cron /etc/crontabs/root
 COPY config/credentials.json /app/credentials.json
 COPY config/token.json /app/token.json
 
-ENTRYPOINT [ "/app/entrypoint.sh" ]
+CMD [ "/app/calimage" ]
