--- calendar/entrypoint.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-/app/refreshserver && cron -f
