--- calendar/go.work
@@ -1,6 +1,0 @@
-go 1.19
-
-use (
-	./imagegen
-	./refreshserver
-)
