--- calendar/refreshserver/go.mod
@@ -1,3 +1,0 @@
-module refreshserver
-
-go 1.19
